backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ultralytics==8.3.40
torch==2.5.1
torchvision==0.20.1
Pillow==11.1.0
python-multipart==0.0.12
pydantic==2.10.4
httpx==0.28.1
